RAG Chatbot using SQL Agent/requirements.txt
langchain==0.0.331
langchain-core
langchain-community
faiss-cpu
langchain_ollama
python-dotenv
SQLAlchemy
mysql-connector-python
langchain-groq
streamlit



